+++ object.hpp
@@ -16,17 +16,17 @@ namespace logic {
 class Object {
 public:
     explicit Object();
-    explicit Object(long x, long y);
-    explicit Object(long x, long y, long d_x,
-            long d_y);
+    explicit Object(double x, double y);
+    explicit Object(double x, double y, double d_x,
+            double d_y);
     explicit Object(const Object &object);
     Object& operator=(const Object &object);
     virtual ~Object();
 
-    void get_dpos(long *, long *) const;
-    void get_pos(long *, long *) const;
-    void set_dpos(long d_x, long d_y);
-    void set_pos(long x, long y);
+    void get_dpos(double *, double *) const;
+    void get_pos(double *, double *) const;
+    void set_dpos(double d_x, double d_y);
+    void set_pos(double x, double y);
 
     double distance(const Object& object) const;
     double speed_diff(const Object& object) const;
